requirements.txt
asgiref==3.8.1
Django==4.2.13
django-allauth==0.63.6
sqlparse==0.5.0
